requirements.txt
streamlit
streamlit-chat
python-chess
openai
python-dotenv
Pillow
faiss-cpu
sentence-transformers
torch
